Remove implicit cast from float to Coord and float to Size Add a new Abestract class to prevent to call to many times the Coord(Vector<float, 3> constructior and Size(Vector<float,3>);
Remove compilation problem cretead by these changes in the rest of the code. 
TODO : Add a #pragma in VectorCast.h to prevent warnings during compilation.
           Bench performance compate to previous implementation.

git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@2227 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-ogl/src/GlCPULODCalculator.cpp
@@ -57,7 +57,7 @@ namespace tlp {
 
       Coord eye;
       if(camera->is3D()) {
-        eye=camera->getEyes() + ( camera->getEyes() -camera->getCenter() ) / camera->getZoomFactor();
+        eye=camera->getEyes() + ( camera->getEyes() - camera->getCenter() ) / (float)camera->getZoomFactor();
         computeFor3DCamera(&simpleBoundingBoxVector[i],&nodesBoundingBoxVector[i],&edgesBoundingBoxVector[i],
             &(simpleResultVector.back()),&(nodesResultVector.back()),&(edgesResultVector.back()),
             eye,transformMatrix,globalViewport,currentViewport);
